# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EDCS API",
        "description": "The Elite Dangerous Companion Suite API provides galaxy data from Elite Dangerous to various consumers, including the EDCS frontend at https://edcs.app. Data is sourced from EDSM (Elite Dangerous Star Map), EDDN (Elite Dangerous Data Network), and the Frontier Companion API (CAPI).",
        "contact": {
            "name": "EDCS",
            "url": "https://edcs.app"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "EDCS API Server"
        }
    ],
    "paths": {
...
        "/systems/search/route": {
            "get": {
                "tags": [
                    "System Search"
                ],
                "summary": "Find the shortest jump route between two systems",
                "description": "Computes the shortest route between two systems within the given jump range. Returns an ordered list of waypoints with per-hop and cumulative distances. Results are cached for 24 hours.",
                "operationId": "6fb108aa18dc9a4d4fcb6f835715cb33",
                "parameters": [
                    {
                        "name": "from",
                        "in": "query",
                        "description": "Origin system slug ({id64}-{name})",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "8216113749-maia"
                        }
                    },
                    {
                        "name": "to",
                        "in": "query",
                        "description": "Destination system slug ({id64}-{name})",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "670685668665-pleiades-sector-ag-n-b7-0"
                        }
                    },
                    {
                        "name": "ly",
                        "in": "query",
                        "description": "Maximum jump range in light years (min: 1, max: 500)",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 40
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ordered list of route waypoints",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/SystemRouteWaypoint"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No route found within the given jump range"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SystemCoords": {
                "properties": {
                    "x": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "y": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    },
                    "z": {
                        "type": "number",
                        "format": "float",
                        "example": 0
                    }
                },
                "type": "object"
            },
...
            "SystemRouteWaypoint": {
                "properties": {
                    "jump": {
                        "type": "integer",
                        "example": 0
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id64": {
                        "type": "integer",
                        "example": 8216113749
                    },
                    "name": {
                        "type": "string",
                        "example": "Maia"
                    },
                    "coords": {
                        "$ref": "#/components/schemas/SystemCoords"
                    },
                    "slug": {
                        "type": "string",
                        "example": "8216113749-maia"
                    },
                    "distance": {
                        "type": "number",
                        "format": "float",
                        "example": 38.2
                    },
                    "total_distance": {
                        "type": "number",
                        "format": "float",
                        "example": 38.2
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum bearer token. Obtain via POST /auth/login or the Frontier SSO flow.",
                "scheme": "bearer",
                "bearerFormat": "Token"
            }
        }
    },
...
}